Clean up strings related to process management
The most important is the thing which gets broadcast in the
"disconnected" signal. This is what gets shown to the user, so it should
really show something which makes it obvious that the connection is now
down. There's probably no reason to mention the fact that we're using a
pipe-based connection to some process.

The other status messages which are handled in the handleStateChanged()
go directly to the error log, so it makes sense to indicate that a
process died. OTOH the QProcess class name is just an internal
abstraction, so let's stop leaking that one, too.

Change-Id: Ief52e93b273d9f12cd9eef3671c575083ccd6b6a