Restore informing about reconnections
This got removed in 7cab7cc3803302bdb21bfd27ee649ddeb981fd2e, and I
suspect that it was an accident because that patch essentially left
behind a useless empty slot.

I also cleaned this up a bit because I really prefer this new syntax --
and it was worth a lambda because it was so trivial, anyway.

Change-Id: I90c92d2ccf9d06822961998791d88c17fd7d97ac