Model: Refactoring: keep track of what was saved into MessageDataBundle
This is a *ton* of boilerplate code, but the motivation behind this is
to get rid of that four explicit bools which handle what exactly was
already stored. Right now, the problem is that these bits live within
the context which processes a single FETCH response, so it's possible
that we fail to set appropriate state when the data are split into
several FETCH responses.

Change-Id: Ic0ba5949d841da715d1bd0d84716844f21108c80