Fix "fetched" vs. "child nodes available" on a TreeItemMessage
Separate the "message is fetched" from "envelope is available" and
"child nodes are available", and make sure we do not accidentally mark
the current message as fully fetched while setting its children.

The Crypto code previously took the message's validity for granted, but
now that we're using more fine-grained approach, let's do it properly.

Change-Id: Ib713f349620f6df0a8772dae6f20899e9b197d23