Model: use MVC signals for message metadata loading
We are inserting tree items, so we should absolutely emit
rowsInserted().

The dance with the fetchStatus is a bit annoying but needed because
the QAIM's protected methods such as beginInsertRows() are happy to call
rowCount() with calls our fetch(), which leads to nasty surprises such
as triggering of askForMsgMetadata() about different messages due to
preload, etc.

Change-Id: Ief621215ee2fc2b3a8cd7f99fa118b3aebc2993d