Crypto MessageModel: fix working with not-yet-loaded messages
The Cryptography::MessageModel failed to account for the use case where
it's created and tied to a message which is not fetched() yet. In that
situation, it is wrong to assume that the message structure is already
available. One should wait for a proper rowsInsrted() and restart the
insertion process at that point.

This *partially* fixes the problem of a message not appearing in the
MessageView when clicked before it was fully loaded. Now it "appears",
but without headers :). That's progress.

Change-Id: I4c4116822c7e9b2b4781c1a79f27f95bed76d6d8