Improve MessageWidget's operation
Fix requesting of messages which haven't finished loading yet. This got
broken "somewhere during the MessageModel introduction".

Fix updates of tags -- the backstory is the same as above.

Switch to a stacked layout, so that we do not have to hide and show
widgets like monkeys, all by hand. That just isn't fun.

Change-Id: Ib8343b949a38eeb5d876198dcc3bb7c4952e8ffe
BUG: 360273