Verify S/MIME (CMS) signatures
Everything still goes through the GpgME, so the changes are just some
infrastructure work and also some fixes for reporting of the "key IDs",
matching the key UIDs, and dumping the trust chain.

There are no tests; in a perfect world, it would be terrific to have
tests set up a custom CA and create some certs, similar to how we
already handle OpenPGP. Patches are welcome :).

Change-Id: I63541741d3a201260b13aa5e99ded396099314d3