Do not leave dangling pointers behind when replacing items
When the MessagePart replacement happens through a registered
PartReplacer, the original ProxyMessagePart is passed to the replacer's
factory method as a unique_ptr, which means that it will go out of scope
if the new part doesn't choose to save a reference for future. This is
safe because the pointer has not been announced anywhere through the MVC
API at this point, so it's just an internal implementation detail -- or
at least that was the idea.

However, it turned out that the proxying parts actually registered
themselves within the model's translation layer (which is a good and
required thing), but their destructors failed to remove this reference.
This patch fixes that.

The destructors are a bit careful, they are not comparing based on the
key within the map because I have no clue if QPersistentModelIndex'
operator< still works even after a possible index invalidation. This
means that for the sake of correctness, we simply remove all these refs
which point to an object that is being destroyed.

There's no associated autotest for this thing, unfortunately. I
triggered this by making a naive change for encrypted S/MIME messages.
In theory, I could probably simply write a trivial replacer doing just
this, but I'm a lazy guy.

Change-Id: I13243603e286013e84fe5c95287fc0fd6a58d174