Composer: generate stable message content
There was an annoying bug that the Message-Id and the string used for
MIME part boundaries was regenerated from random source when multiple
calls to writeCommonMessageBeginning were made. That's wrong because the
code was making legitimate calls to that function when, e.g., the IMAP
server supported CATENATE for saving of messages, and the SMTP server
did not support BURL (or the IMAP server didn't support URLAUTH, for
that matter).

This patch ensures that these random strings are generated on first use,
and then remembered. As a result of that, the content in the Sent folder
now matches what actually got sent over an MSA.

Many thanks to Erik Quaeghebeur for catching this -- and sorry that you
had to go all the way to the mlmmj upstream in troubleshooting this.

Change-Id: I932405c7d3f5be7f82c94041c2c2ff4ca8586e01