Simplification: range based for instead of a std::for_each
This code was originally erasing some of these members which means that
a range based for couldn't be used. However, now it doesn't erase (as
we're in the destructor anyway), so a range based for can be used. I
prefer it because it's shorter and because it doesn't require a lambda.

Change-Id: I95e964d62fdd7cf863ba352001a3d252584eca89