consistently use escaped double quotes in html strings
In the source code, when html attributes appeared in strings,
both single quotes and escaped double quotes were used. For
example, "<img src='...'/>" and "<img src=\"...\"/>". This
patch makes the usage consistent, meaning that single quotes
usage was eliminated. A reason to use double quotes instead
of single quotes is that for strings subject to replacement
(with html-escaped content) it is unsafe to use the latter,
as single quotes are not escaped by Qt's toHtmlEscaped.

Change-Id: I418ea4f973009d7ba00204e4a93b7dd2be2ce19d