FindBar: support operating on plain QWebView instances, too
The FindBar implements not just the generic "find bar" functionality,
but it also has features for driving our EmbeddedWebView in an
appropriate way -- mostly related to the automatic scrolling. This patch
makes sure that the FindBar continues to work even in cases when the
associated widget is a plain QWebView. This was the path of a least
resistance.

Change-Id: I7007e2934eae9dec5274f27bd607e3bba0af0542