Implement Ctrl+F in the message headers view
This is different from the "message source view" because the message
source view has historically used a different implementation of the data
retrieval with a FullMessageCombiner.

I'm more and more inclined towards a solution which moves this part
joining all the way into the IMAP Model, with a special part column
offset (similar to the existing .HEADER modifier, for example). That
would make this code much more straightforward, and it would also
eliminate the need for a FullMessageCombiner and the associated extra
code paths from several places.

The change from SimplePartWidget to a QWebView is semi-intentional; a
SimplePartWidget requires special embedding, and we probably do not want
that. It would also be nice to implement both this widget and the "show
headers" in terms of an EmbeddedWebView, or to split the safety feature
tweaks of the EmbeddedWebView into an extra class and use that. That's
also work for another day, though.

Change-Id: I769a01ba48ff5337333e9e1bef3c43cd10e8c9ed