Save the GUI settings in one-at-a-time mode, too
I guess that this comment was originally meant to explain that it
doesn't make sense to store the splitter's positions when the
one-at-a-time mode is active. However, there are other settings which
absolutely do make sense to be stored. Also, storing the splitters'
stuff does not cause any troubles, which is the reason why this patch is
so straightforward.

Thanks to CcxCZ on IRC for reporting this.

Change-Id: I9e9b36c2910ea3d61fb9f8ef88b9ffce34615ecc