Use asynchronous callbacks when prompting for passwords
This means that we are now safe against the usual QDialog::exec() race
condition.

At first, I wanted to do this in a C++11-ish way with two std::function
arguments, one called on success and the other on failure, but doing
that would require some add-on fancy decorator which stops the signal
forwarding once the pointed-to QObject is destroyed, so I chose this
way. A disadvantage of that is that one has to manually ensure that
rejected() and gotPassword() are connected, which might be error prone.
An alternative would be to leverage bits from QtPrivate for signal
forwarding, but meh.

Change-Id: I400718758f4cacdb735e7b820759ba1fb93fc3c1