Requse password dialog on network reconnect cycles
When Trojita was started in its default settings (go online when
possible, do not remember passwords, watch for network session state
changes), no password was available and the network got cycled, a new
password prompt was opened. This patch ensures that if a password
request is pending, no additional prompts are enqueued.

This is implemented in the IMAP Model, so it isn't GUI-specific.

BUG: 362477
Change-Id: I4b14f06bc70b3580f16b0bfa4e47fcdb98b9d353