GUI: Update zoom mode for all parts in sync
This is similar to that Ctrl-F shortcut handling -- if we have more than
a single part in an e-mail, multiple actions with the same shortcut are
going to be active, which is a Bad Thing™.

Let's handle this centrally, then. Provisions are still in place for a
possibility of no enclosing MessageView.

Change-Id: I778f479877b830824e0556220ae8d3a0d50819c7