IMAP: Implement RFC 7888's LITERAL- extension
This is a recent addition to the ever-growing family of IMAP RFCs. Its
aim is to solve a problem with LITERAL+ which allows and encourages IMAP
clients to send big chunks of data without asking the server for an OK
first. It turns out that this sometimes lead to wasted bandwidth.
LITERAL- on the other hand only allows this shortcut to happen for data
of up to 4096 kB.

As I expect some IMAP servers to start offering the LITERAL- extension
eventually, it is a good idea to be ready for this. If we didn't
implement LITERAL-, we would be stuck with the synchronizing literals
and therefore extra delays due to the network's RTT on those upcoming
servers at some point in future.

I'm so looking forward to servers which misimplement the RFC and start
rejecting based on ">= 4096" rather than "> 4096" :).

Change-Id: Ic0405dbd680d50f608a4c1644f07785c86b516e3