SVG Breeze icons
This is a great opportunity to, well, change the UI to something which
looks modern. Using a reasonable icon scheme and especially the choice
of its default shape is a nice topic for bikeshedding, so I'm simply
going to use that piece which I get on my laptop by default anyway.

BUG: 360689
Change-Id: I74a0e36d89fde0dcfb798a0718c8dd256811c2aa