Fix tooltips in QtWebKit
Something™ triggers an event that makes qtooltip feel required to hide.
Next event cycle is by far not sufficient, so hacko mui uglydo would be
this thing.

But this is a "bug" in QToolTip or QWebPage/View or in Qt's event
handling (ie. Qt5 is in poor shape anyway and getting worse in every
update)

No, it's afaics not the hideText() call ;-)

jkt: It's actually caused by WebKit trying to handle the popups
internally and by an unfortunate order in which these events are fired.

This only works on Qt 5.4+ because that's when qtbase got its commit
e2df05f120ea76f7fb777926f5c00163dba4a92b. In other words, the bug
remains unfixed for Qt 5.2 and 5.3. I can live with that.

BUG: 363783
Change-Id: I03373a86695a8813cef70a964a2168e78a311368