IMAP: do not misinterpret pending grandchild mailboxes
Under certain circumstances where the code triggered several LIST
commands at once, the LIST interpretation was a bit too eager and simply
looked at all mailboxes whose name matched a correct prefix. This worked
as long as we had a guarantee that mailbox listing for "A" and "A.B" was
never active at any given time, but this guarantee doesn't really hold
anymore.

It could be argued that it's actually a bug to ask for parent's and
a child's children at the same time (because the child data will get
thrown away anyway due to our enforced re-creation of all these
TreeItemMailbox instances), but let's solve this in a pretty quick and
just-a-bit-dirty way. This patch simply ensures that we don't assume
that these grand-children are actually children of the current parent.

BUG: 364314
Change-Id: I37570a6a5eab3eb15dcdbfa29922cb1b8dd79a7a