GUI: About box: talk about Qt
This adds the "Built with Qt" logo, hopefully according to the Qt's
trademark policies outlined in [1].

I don't have any real reason for adding this except being a little bit
bored, and willing to advertise our usage of Qt. So I guess that their
blogpost worked, it managed to get my interest :).

The dimensions and alignments and spacer behavior are altered so that
the end result looks a bit more nicely altogether in my opinion.

[1] http://blog.qt.io/blog/2016/06/20/introducing-new-built-with-qt-logo/

Change-Id: I34c0442ddc95f1adce58bbfda877132fea0d91d9