IMAP: Accept more spaces when reading untagged responses
The IMAP standard is 100% clear on this one, but apparently there are
servers out there which get this thing wrong and send extra spaces every
now and then -- a "DevMail Gateway" is one of these, according to the
linked bug. That behavior is incorrect and in violation of RFC3501, but
whatever, we've added similar incompatibility quirks before and I'm
afraid that we'll have to keep adding them in future, too.

Let's see if this is enough to work with the above mentioned server
implementation.

BUG: 365149
Change-Id: Iff89b82ca3ca4be719076e3ee81176b8fc7d287c