Fix segfault at exit related to the Icon cache clearing
Commit 8bb25e28b4173aaa11aa9223f58e3f85e4293d29 fixed it partially. But it
still crashes with XdgIconLoader, the LXQt IconLoader.
(https://github.com/lxde/libqtxdg). XdgIconLoadet it's a tweaked Qt own
QIconLoader.
Q_GLOBAL_STATIC is the right tool for this job. It also clears the cache
when the application is closed with the Titlebar Close button, which
wasn't being done.

Crash bt:
    at ../xdgiconloader/xdgiconloader_p.h:96
    at ../xdgiconloader/xdgiconloader_p.h:96
    at /usr/include/qt/QtCore/qalgorithms.h:323
    at /usr/include/qt/QtCore/qalgorithms.h:331
    at ../xdgiconloader/xdgiconloader.cpp:441
    at ../xdgiconloader/xdgiconloader.cpp:440
    __in_chrg=<optimized out>) at /usr/include/qt/QtCore/qhash.h:147
   from /usr/lib/libQt5Core.so.5

Change-Id: I004c40d92ef354b3c035b86456024a9e155c7e92