Fix assert on saving settings
Upon saving the IMAP settings, Gui::Window::nukeModels is called which
in turn calls MailBoxTreeView::setModel(nullptr) which asserts in
Imap::Mailbox::MailboxFinder constructor. This was introduced in commit
6fd4529e23f127308a3e178b1c3089aba9dcabdf (/me waves).

Given that the MailBoxTreeView already handles a nullptr
m_mailboxFinder, let's just take an advantage of that.

I guess that we should do some real-world UI testing in order to catch
problems like this one.

Change-Id: I171c2750edf4dfa80f2d6a08fcaba3886bad803e