Plugins: do not assert on plugins duplicated to several directories
As Caspar points out, the code asserts when a same-named plugin is found
both in the systemwide directory and in our current build directory.

This change makes sure that only one "instance" of a same-named file is
attempted for loading. This check happens early in the path and because
appdir-local plugins are actually attempted to load earlier than the
rest, it Should Work™.

BUG: 367607
Change-Id: Id0bffffb508a03708c1cbe8d43b8685c236a3292