tests: Remove Linux-specific includes for dlsym
The dlsym(RTLD_NEXT) appears to be reasonably portable (it's POSIX,
except the minor RTLD_NEXT and _GNU_SOURCE on glibc), so let's use this
thing to fix a build failure on OpenBSD.

We are doing this because GnuPG 2.1+ is still not really that
mainstream and because faking the random generator just for the tests at
all platforms simplifies the test matrix. Alternatives such as GnuPG
version detection are IMHO too fancy.

BUG: 367601
Change-Id: Icf0001a20c6e2901a88b65a705059d687951b1fb