net: Do not signal QProcess progress which might be misleading
The network detection code assumed that once a connection reaches the
CONN_STATE_CONNECTED_PRETLS_PRECAPS or CONN_STATE_SSL_HANDSHAKE states,
the network can be safely considered as "working", and the network
reconnect timer therefore resets to its shortest interval.

This assumption wasn't playing well along with QProcess that launching
an SSH connection. The ProcessSocket::handleStateChanged was a bit too
optimistic in emitting CONN_STATE_CONNECTED_PRETLS_PRECAPS to actually
mean "yeah, the process itself has started, but it might still be doing
$stuff and network might still be unavailable".

A correct fix probably depends on one's perspective. One could argue
that the QProcess backend should use CONN_STATE_HOST_LOOKUP ->
CONN_STATE_CONNECTING, and only report CONN_STATE_CONNECTED once a first
byte of output appears, for example. However, this smells a but tricky
(there's a possible race in there, and no other backend notifies about
incoming data prior to CONN_STATE_CONNECTED as far as I know). So in the
end, this patch changes the timer to only treat a 100% succeeding
connection as "ok, connected".

This has a potential side effect with password prompts that are ignored
for long intervals -- but maybe it's actually correct to stop prompting
for passwords several times on top of each other :).

Change-Id: I07dab3ab852823833b933ce3a436643312d8fed9