Do not enable sanitizers for the /dev/urandom fake wrapper
That library is injected into the gpg binary, and as such, our build
preferences do not really make much sense for that. And after all, ASAN
doesn't really work in libraries when the application doesn't use ASAN.

Thanks to Hanno Böck and his QtCon talk for reminding me about ASAN.

On a side note, I wonder if it's even possible to do cross-compiling
with this wrapper.

It turned out that disabling automoc is enough -- this is correctly
detected as a C file, and therefore our CXXFLAGS take no effect, which
is wonderful.

Change-Id: I545f5fed44ff52ddfa61514cb6261c24c24d5d6d