AbstractCache shouldn't really be a QObject
This is motivated by some refactoring within the test suite; I'm trying
to be consistent with use of smart pointers, and AbstractCache being a
QObject class interferes with it. It turned out that this QObjectness
was only used for the error signal, and that thing can be very easily
represented by a std::function and a couple of lambdas these days.

It was a bit too intrusive, though :)

Change-Id: I1c72440021ae2507c8d8444683b35e817c94e530