Fix Socket's ownership
The Parser should really own the Socket that it's working on at all
times.  We have been already sort-of doing this through that call to
sock->deleteLater() from within ~Parser(), but the catch with this thing
is that the event loop might not be running anymore at that point (as
far as I understand the root caluse behind the LSAN-reported leaks).

Change-Id: I33fdc1947e08957f12cbb9bbaa0b8ef24c5ecd6d