tests: Isolate the tests from the D-BUS env
This started when ASAN-ing a test for MsgPartNetAccessManager. Turned
out that that thing loads real platform-specific modules, which is not
something that we want.

Let's ensure that our tests are sufficiently independent here (and let's
wait for a first test some time in future which happens to talk to a
mocked systemwide service...).

Change-Id: Ifb28383a99833e1ae7e8ff9c7572430840c47569