Tweak date format of several-months-old e-mails
While reading some e-mails from an RSS feed, I got confused a little bit
by messages saying "23 Nov". I think that it's OK to assume that
December will be unambiguous when it's January of February, but showing
previous year's Oct/Nov stuff in September is probably not that
intuitive.

I wonder if this is just me, or if others perceive it in this way,
too...

Change-Id: I9134f11c43b301daa3dba596df62710769e3ffd9