tests: Skip QtWebKit tests when building with ASAN
So yeah, this is evil, but without this patch, I'm getting a funny
looking failure from ASAN's stack-buffer-overflow from within qtwebkit's
JSC.

Given that QtWebKit is unsupported an dead for all practical purposes, I
suppose that there's no chance of an actual fix hitting their tree. I
could probably instrument the CI builds to skip the relevant ASAN checks
for QtWebKit, but it isn't as easy as just not using -fsanitize=address
because we still have to link with the appropriate libraries because the
rest of Qt is ASAN-enabled.

So this change just prevents the offending webkit-using code to execute;
the idea is that the test won't fail, and we will be able to focus on
other ASAN failures in the rest of the build.

Change-Id: I6fefdae1cdd7ed730a2742ff95e9547933f6520e