Fix GPG test
At least GnuPG, when run with both --fingerprint and --with-colons will output
the fingerprint of the public key and the subkey. The command below will thus
output two fingerprints and the delete command will fail because it cannot
find the second fingerprint (which is a subkey and not a public key). It seems
that a safe approach is to consider only the first fingerprint line.

Change-Id: I4c6aa79e4473bca146aa087728a5652d91586a81