GUI: Reduce indirect QDialog::exec calls
QDialog::exec is dangerous because it re-enters the event loop. We've
got a report of a crash that can be traced back to
MainWindow::cacheError which led to further event processing up to
Gui::MainWindow::slotPluginsChanged which tried to manipulate QActins
which weren't really instantiated at that time.

Reported on Haiku (!) by Zoltán Mizsei ("miqlas" on IRC).

Change-Id: Ibd24040c67341961b69cd058253fceadf5a6ea2a