GUI: Next/Previous mail actions should work from all contexts
v2: The QAction binding was correct, the bug was in our logic. When no
items were selected, the move would default to a no-op.

We are not going to the very first item in a mailbox because doing that
might be a bit counter-intuitive on huge mailboxes. This patch will
resume the navigation at the top of the current view, depending on where
the user scrolled the viewport. Chances are that this is going to be
near the desired position.

Thanks to Sérgio Martins for reporting this.

Change-Id: I89fde26bbd5ff4ae445410eff0d22f39ca996ee1