Work around MSVC2015's alignof()
It seems [1] that MSVC doesn't support calling alignof() on abstract
classes, but one has to make a detour through a reference. Documentation
says that the alignment of a reference type is the same as that of the
referenced type, so we should be safe.

Thanks to Sérgio Martins for reporting this.

[1] https://www.reddit.com/r/cpp_questions/comments/45v1oh/alignment_of_abstract_class/

Change-Id: Ia4d9b4f845bb6f65a425f3d422f362f45b9e2214