GUI: Add support for setting per-account icons
I have several Trojita windows open at any given time, one for each
account that I use. Some of my virtual desktops therefore get a bit too
crowded, and it is hard to distinguish between the top-level windows
which belong to different e-mail accounts. This patch fixes it.

When we move to supporting several accounts within a single instance in
future, I do not expect any breakage caused by this change. The icon is
being tracked in the IMAP account level (and not per-identity, for
example). This will make it easy to e.g. add icons to the mailbox tree
with pretty decorators for each account, etc.

There's no fancy UI yet, patches are kindly welcome. A naive file-picker
is not enough, though, because there are (IMHO) valid use cases for
setting e.g. an RSS-like icon for that profile which is used exclusively
for reading of news feeds sent to e-mails, etc. In those situations,
specifying just "application-rss+xml" and relying on the icon engine to
do the trick might be a reasonable approach. Or maybe not, because
Plasma uses a bigger icon in the taskbar while kwin is happy with a
smaller sized one, and these are visually different in current
breeze-icons.

Change-Id: I021794cf4e5c21997a6da9c6091cc5b5b24a1e81