GUI: Gray out systray icon when not actually connected
The connection state tracking is tricky, there's a window where the
IMAP Model's networkPolicy() reports the intended policy, but where the
actual network connection has not succeeded yet. This can happen for
example when the system as a whole believes that we /might/ still be
online, but where the network connection is already dead and Trojita
happens to be reconnecting or waiting for a timeout.

This change ensures that the systray icon only reports success where the
connection is alive and OK.

This patch might be a bit too intrusive because the NetworkWatcher's
status is also used in part loading to control whether it makes sense to
ask for certain parts. However, because we always invalidate indexes at
reconnect time anyway, this is a safe thing (as in "won't introduce
visual regressions) because if we ever reconnect, the part widgets will
loose their state anyway.

Change-Id: I2e90f237ceff5fe2e1635f41456eaf793041535e