Gui: use QaimDfsIterator when checking for previous/next unread messages
Now that we have support for real iterators, there's no point in
duplicating that work here.

The code is still more complex than I would like to see because of the
circular wrapping. Ideally, another extra iterator adaptor should be
done for this thing, but that will have to wait for another time.

Change-Id: I6372552f0fbaf6f7c532fd313c166b32b5d1b042