Refactor & test moving to next/previous unread messages
Now that we have our fancy iterators, make sure that they work and are
automatically tested. The interesting corner cases was wrapping the
iterator around the begin/end of the entire QAIM as well as doing "the
right thing" (whatever it is) when dealing with reverse iterators and
the current item.

Change-Id: If31996827a4a287142f75c7d70a7bd3e61c8b7fc