Gui: Prevent flickering of button text when typing into Composer
Without this patch, constant changes can be amplified by Plasma's
platform integration plugin which unconditionally adds accel keys. That
leads to visual flickering (which is annoying) and also to increased
power consumption (which is probably negligible compared to the amount
of energy wasted by our CI even when we integrate that of all our users,
but bah. Yeah, I just returned from a sauna & beer session.).

Change-Id: Ie818015c649564e695c8ea20c780d6d57dfd2e2a