IMAP: Retry by fetching the raw part on [UNKNOWN-CTE]
An IMAP server supporting the BINARY extension can return UNKNOWN-CTE
when decoding of a message part failed. In particular, Dovecot simply
says * NO [UNKNOWN-CTE] for these parts.

Let's go back to local decoding for those parts which actually failed to
load upon seeing the [UNKNOWN-CTE] response code. Try to check it
regardless if it came in an OK or NO response for the FETCH command.

BUG: 341982
Change-Id: Ie782f0c89879cc270234e7f057ab3fe38364a12d