Allow accessing network connectivity info via the MVC API
I have mixed feelings about this one -- on one hand, it is "ugly"
because this information is not really tied to any specific index and
any changes in it will not result in a dataChanged() notification for
obvious performance reasons. However, on the other hand having access to
that piece of information enables looser coupling between components --
it is no longer needed to access the underlying Imap::Mailbox::Model,
and it will be used in the next commit when we might need access to that
piece of information from inside a context where there is only
Crypto::MessageModel.

Change-Id: Ib06eadd3db204be0ccfac6f98d18457b51c7fc16