Import a version of reverse_iterator which supports stashing iterators
A regular std::reverse_iterator requires a BidirectionalIterator, but
the QaimDfsIterator is not even a ForwardIterator, just an InputIterator
(it's a generator, and one cannot access the element returned as a
"reference" after the iterator has been destroyed).

This implementation follows the C++11 wording (which got reverted in
C++14 due to defect report 2360 [1]) and -- at a cost of an extra member
and a non-const operator*() -- supports reversing also those "stashing
iterators".

[1] https://timsong-cpp.github.io/lwg-issues/2360

Change-Id: I1aa5f454546ee5bf2b87dd11a48aa83df026350f