GUI: Do not keep reusing wrong password
Since commit 76648e4248, Trojitá has started forgetting to forget a
cached password which proved to be wrong. When a password was provided
by some plugin, the UI would never ask the user for one after a login
failure. That's suboptimal, especially considering that the code would
just happily loop and issue many login requests using the same
credentials over and over again.

Thanks to Marc Lambrecht (whytrytofly on IRC) for reporting this.

Change-Id: I1920176ad71910bea6e8c83033e444dc01c076bf
BUG: 375822