GUI: Extracting e-mail addresses from a message body
Right now, we delegate link copying to WebKit, which means that it will
use URLs like "mailto:foo@example.org". That is suboptimal from the
user's point of view; they typically just "want to have that e-mail
address".

This patch ensures that we do this fancy thing when the URL is really a
simple one -- pointing to exactly one recipient. As a fallback option,
all other cases are handled by the existing code which just copies the
URL as-is.

Please note that this does not fix the issue that the linked bugreport
complained about -- the header widget with e-mail addresses is a QLabel,
so we will have to do a slightly different dance for that.

Change-Id: I930f4eecd291af643c736701849d4c92fc8fc43d
CCBUG: 374830