GUI: Clipboard copying of e-mail addresses from the header view
Previously, the context menu offered some options for accessing the
QLabel (this class is actually just a pretty QLabel)'s text, for
selecting all of it, and copying stuff.

The most important, IMHO, is to have access to the actual e-mail address
(as requested by the associated bugreport), and to the human readable
text for the sake of completeness.

The code is still a tad inconsistent because the List-Post header for
indicating an active mailing list does not currently use the
OneEnvelopeAddress class. Mailing list addresses will therefore remain
unaffected by this patch; their copy action is still going to copy a
complete mailto: URL.

BUG: 374830
Change-Id: I4e515045d0241e4c1064389aa777f58e72c7c999