Enable monitoring of additional mailboxes in the systray
There are situations where it makes sense to show an aggrgetaed number
for all messages in all folders (use case: an RSS feed reader living in
an extra profile). The warnings are there to discourage users from
hitting this in situations where they have a huge set of shared folders
from 2k people in the same company, etc.

Now that we have that fancy QAIM tree iterator, the code is super-easy.
I like this :).

Change-Id: I4a9b4771c56324c46631abe73e4b9d21a4045fdd