Add shortcut for focusing mailbox tree and message list
There was no reproducible way to focus the widgets via keyboard
while those are important elements of navigation for the user.
The focus can be "lost" in the message widget and it is not possible
to predict how many number of tab hits needed until either widget can
get it.

Change-Id: Icb84c85c37c277867e516d9820ad6f857404ded7