Move cursor to the next message after current one disappears
A big inconvience for keyboard usability if the cursor forgets last
position when the current message disappeared from message list. It
also prevents next/previous message actions to function.
This patch using an unobtrusive approach to this problem, moves the
cursor to the next message but does not activate it. The user can
activate with return key or can navigate towards. This way it does not
interfere with automatic read marking nor habits of users.

CCBUG: 360091
CCBUG: 321387
Change-Id: Ic823c2c9d0bba9de65d01fb0e72b91093f233f7e