Propagate color of selected messages too
Selected messages has lost their color which is a UI defect. This patch
changes this behaviour to how Thunderbird works: show the color as row
background when the list widget is active and simply keep the color of
the text when inactive.
The settings page also follows the same principle, so the user can get
some clue how it will look like in action before applying.

Change-Id: I4dd4374ba728fe0c4d9dcc8a8e9eaec6f9eaabe4