Do not autoexpand childless messages in the message list
Users familiar with Qt applications are expecting that pressing left
arrow key on a childless message immediately navigates up to the parent
in the tree but actually it's not happened at the first keystroke in a
newly expanded thread and it confuses the user. This patch is a fix for
this keyboard behaviour.

Reproducing:
1. Expand a thread
2. Navigate to a childless message in the thread
3. Press left arrow key (once)
Actual results: Nothing happens. Pressing again jumps to the parent.
Expected behaviour: Jump to the parent at the first keystroke.

Change-Id: Id0bc71df233ca934312ac105b248f280112d3671