Fix "Key sequence 'Ctrl+F' is ambiguous" error message
If there are several SimplePartWidget instances which are visible at
once, then it is ambiguous which one is going to handle the Ctrl-F
shortcut. This patch fixes it by routing the Ctrl-Fs via the
MessageView. All multipart wrappers are changed so that they offer this
action to all child widgets in a round-robin fashion until one of them
"swallows" it.

Change-Id: I053a0a3aa676d7c4443af852eeaf86e22c1af8c9