UI support for message bouncing ("open in composer")
I started contributing to some projects which follow the LKML-style of
development with patches being sent via e-mail. At the same time, some
of these projects do not like MIME for some reason and have a preference
for using `git-send-email` for patch preparation.

Copy-pasting patches is a no-go because of format=flowed. This is by
design -- if a project is so insisting about e-mail structure, manual
tweaking within a message composer which was never designed to be a code
editor is probably not a good idea.

I do not really want to fight the windmills, so let's support a workflow
where Trojita is used just for the actual e-mail submission. For a
full-stack solution, I still have to upload the "patches" in form of
.eml messages somehow. Right now, drag-and-drop from a GUI application
is the only way.

Also, this change is *not* a full-blown "let's preload the composer with
an arbitrary message". That's a whole different kettle of fish because
we do not currently have any writing facilities in our MIME tools.
Producing arbitrary MIME messages is something which therefore remains
out-of-scope.

Change-Id: I360cda5d430e8a442b364cb200c01d94c17d8dbc