GUI: Show when we cannot send yet/anymore
If an attachment was not available (either "yet" due to a slow network,
or "not anymore" due to e.g. a deleted item), the UI would refuse to
prepare for sending, but there was no visible indication about the
culprit to the user.

This change doesn't make it 100% foolproof, but at least it shows that
something was wrong and that the button hit was not ignored.

Change-Id: I67776902dfaf09410fd84e1bb33c46b1174dbdd0